Keep todo behind auth with owned rows

+++ todo/views.py
@@ -1,11 +1,12 @@
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.mixins import LoginRequiredMixin
 from django.shortcuts import render, redirect, get_object_or_404
 from django.http import HttpResponse
 from django.views import View
 from .models import Todo
-import json
 
 
-class TodoView(View):
+class TodoView(LoginRequiredMixin, View):
     template_name = 'todo.html'
 
     def get(self, request):
@@ -27,7 +28,7 @@ class TodoView(View):
                 context = {'todo': todoitem}
                 return render(request, self.template_name, context)
 
-        context = {"todos": Todo.objects.all().order_by('-created_at')}
+        context = {"todos": Todo.objects.filter(user=request.user).order_by('-created_at')}
         return render(request, self.template_name, context)
 
     def post(self, request):
@@ -35,11 +36,12 @@ class TodoView(View):
         if not title:
             request.session['error'] = 'input is required'
             return redirect('todo')
-        todoitem = Todo.objects.create(title=title)
+        todoitem = Todo.objects.create(title=title, user=request.user)
         request.session['todo_id'] = todoitem.id
         return redirect('todo')
 
 
+@login_required()
 def mark_done(request, id):
     todoitem = get_object_or_404(Todo, pk=id)
     print(request.POST.get('done'))
@@ -49,7 +51,11 @@ def mark_done(request, id):
     return redirect('todo')
 
 
+@login_required()
 def delete_todo(request, id):
     todoitem = get_object_or_404(Todo, pk=id)
-    todoitem.delete()
+    if todoitem.user == request.user:
+        todoitem.delete()
+    else:
+        return HttpResponse(401)
     return HttpResponse(status=200)
